Move formatting methods into FormattingUtility
