+++ Apps/Admin/Client/Pages/Covid19TreatmentAssessmentPage.razor
@@ -51,7 +51,7 @@
 
 <HgBannerFeedback
     Severity="@Severity.Info"
-    IsEnabled="@(StatusWarning != null)"
+    IsEnabled="@(!string.IsNullOrEmpty(StatusWarning))"
     TResetAction="PatientSupportActions.ResetStateAction"
     Class="mt-3"
     DataTestId="user-banner-feedback-status-warning-message">
