--- Apps/Admin/Client/Utils/CommunicationUtility.cs
@@ -1,39 +1,0 @@
-﻿// -------------------------------------------------------------------------
-//  Copyright © 2019 Province of British Columbia
-//
-//  Licensed under the Apache License, Version 2.0 (the "License");
-//  you may not use this file except in compliance with the License.
-//  You may obtain a copy of the License at
-//
-//  http://www.apache.org/licenses/LICENSE-2.0
-//
-//  Unless required by applicable law or agreed to in writing, software
-//  distributed under the License is distributed on an "AS IS" BASIS,
-//  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-//  See the License for the specific language governing permissions and
-//  limitations under the License.
-// -------------------------------------------------------------------------
-namespace HealthGateway.Admin.Client.Utils
-{
-    using HealthGateway.Common.Data.Constants;
-
-    /// <summary>
-    /// Utilities for interacting with communications.
-    /// </summary>
-    public static class CommunicationUtility
-    {
-        /// <summary>
-        /// Returns the formatted representation of a communication's status.
-        /// </summary>
-        /// <param name="status">Status to be formatted</param>
-        /// <returns>Friendly string value representing status value.</returns>
-        public static string FormatCommunicationStatus(CommunicationStatus status)
-        {
-            return status switch
-            {
-                CommunicationStatus.New => "Publish",
-                _ => status.ToString(),
-            };
-        }
-    }
-}
