+++ Apps/Admin/Client/Utils/FormattingUtility.cs
@@ -15,13 +15,71 @@
 // -------------------------------------------------------------------------
 namespace HealthGateway.Admin.Client.Utils
 {
+    using HealthGateway.Admin.Common.Constants;
     using HealthGateway.Common.Data.Constants;
+    using HealthGateway.Common.Data.Models;
 
     /// <summary>
-    /// Utility for formatting enums and booleans into human-readable strings.
+    /// Utility for formatting enums and booleans into human-readable strings for display.
     /// </summary>
     public static class FormattingUtility
     {
+        /// <summary>
+        /// Formats a broadcast action type for display.
+        /// </summary>
+        /// <param name="actionType">The broadcast action type to format.</param>
+        /// <returns>A string formatted for display.</returns>
+        public static string FormatBroadcastActionType(BroadcastActionType actionType)
+        {
+            return actionType switch
+            {
+                BroadcastActionType.None => "None",
+                BroadcastActionType.InternalLink => "Internal Link",
+                BroadcastActionType.ExternalLink => "External Link",
+                _ => actionType.ToString(),
+            };
+        }
+
+        /// <summary>
+        /// Formats a broadcast's enabled status for display.
+        /// </summary>
+        /// <param name="enabled">The enabled status to format.</param>
+        /// <returns>A string formatted for display.</returns>
+        public static string FormatBroadcastEnabled(bool enabled)
+        {
+            return enabled ? "Publish" : "Draft";
+        }
+
+        /// <summary>
+        /// Formats a communication status for display.
+        /// </summary>
+        /// <param name="status">The communication status to format.</param>
+        /// <returns>A string formatted for display.</returns>
+        public static string FormatCommunicationStatus(CommunicationStatus status)
+        {
+            return status switch
+            {
+                CommunicationStatus.New => "Publish",
+                _ => status.ToString(),
+            };
+        }
+
+        /// <summary>
+        /// Formats a communication type for display.
+        /// </summary>
+        /// <param name="communicationType">The communication type to format.</param>
+        /// <returns>A string formatted for display.</returns>
+        public static string FormatCommunicationType(CommunicationType? communicationType)
+        {
+            return communicationType switch
+            {
+                CommunicationType.Banner => "Public Banner",
+                CommunicationType.InApp => "In-App Banner",
+                CommunicationType.Mobile => "Mobile Communication",
+                _ => string.Empty,
+            };
+        }
+
         /// <summary>
         /// Formats a data source for display.
         /// </summary>
@@ -44,6 +102,53 @@ namespace HealthGateway.Admin.Client.Utils
                 DataSource.SpecialAuthorityRequest => "Special Authority Requests",
                 DataSource.BcCancerScreening => "BC Cancer Screenings",
                 _ => dataSource.ToString(),
+            };
+        }
+
+        /// <summary>
+        /// Formats an agent's identity provider for display.
+        /// </summary>
+        /// <param name="identityProvider">The identity provider to format.</param>
+        /// <returns>A string formatted for display.</returns>
+        public static string FormatKeycloakIdentityProvider(KeycloakIdentityProvider identityProvider)
+        {
+            return identityProvider switch
+            {
+                KeycloakIdentityProvider.Idir => "IDIR",
+                KeycloakIdentityProvider.PhsaAzure => "PHSA",
+                _ => identityProvider.ToString(),
+            };
+        }
+
+        /// <summary>
+        /// Formats a patient status for display.
+        /// </summary>
+        /// <param name="status">The patient status to format.</param>
+        /// <returns>A string formatted for display.</returns>
+        public static string FormatPatientStatus(PatientStatus status)
+        {
+            return status switch
+            {
+                PatientStatus.NotFound => "Patient not found",
+                PatientStatus.Deceased => "Patient is deceased",
+                PatientStatus.NotUser => "Patient is not a user",
+                _ => string.Empty,
+            };
+        }
+
+        /// <summary>
+        /// Formats a patient query type for display.
+        /// </summary>
+        /// <param name="queryType">The query type to format.</param>
+        /// <returns>A string formatted for display.</returns>
+        public static string FormatPatientQueryType(PatientQueryType queryType)
+        {
+            return queryType switch
+            {
+                PatientQueryType.Hdid => "HDID",
+                PatientQueryType.Phn => "PHN",
+                PatientQueryType.Sms => "SMS",
+                _ => queryType.ToString(),
             };
         }
     }
